+++ src/components/spartan/SSidebar/index.ts
@@ -3,4 +3,4 @@ export { default as SSidebarItem } from './SSidebarItem.vue';
 export { default as SSidebarItemGroup } from './SSidebarItemGroup.vue';
 export { default as SSidebarSeparator } from './SSidebarSeparator.vue';
 
-export type { TSidebarEmits, TSidebarProps, TSidebarItemProps, TSidebarItemGroupProps, SSidebarSeparator } from './types';
+export type { TSidebarEmits, TSidebarProps, TSidebarItemProps, TSidebarItemGroupProps, TSidebarSeparatorProps } from './types';