+++ src/expose/plugin.js
@@ -1,8 +1,8 @@
 // import '../../dist/style.css';
-import { withOptions } from 'tailwindcss/plugin';
+import plugin from 'tailwindcss/plugin';
 import tailwindForms from '@tailwindcss/forms';
 
-export default withOptions(
+export default plugin.withOptions(
     function (options = {}) {
         return function (config) {
             tailwindForms().handler(config);